+++ app/src/main/java/com/example/todoapp/model/Category.kt
@@ -7,6 +7,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "category_table")
 data class Category(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "category_id") val id : Int,
+    @ColumnInfo(name = "category_id") val id : Long,
     @ColumnInfo(name = "category_title") var title : String,
-    @ColumnInfo var completedPercent : Int)
+    @ColumnInfo var completedPercent : Float)
